+++ src/spellbot/actions/admin_action.py
@@ -58,7 +58,7 @@ class AdminAction(BaseAction):
         def new_embed() -> Embed:
             embed = Embed(title=f"Configuration for channels in {guild['name']}")
             embed.set_thumbnail(url=self.settings.ICO_URL)
-            embed.color = discord.Color(self.settings.EMBED_COLOR)
+            embed.color = discord.Color(self.settings.INFO_EMBED_COLOR)
             return embed
 
         def update_channel_settings(
@@ -125,7 +125,7 @@ class AdminAction(BaseAction):
         def new_embed() -> Embed:
             embed = Embed(title=f"SpellBot Player Awards for {guild['name']}")
             embed.set_thumbnail(url=self.settings.ICO_URL)
-            embed.color = discord.Color(self.settings.EMBED_COLOR)
+            embed.color = discord.Color(self.settings.INFO_EMBED_COLOR)
             return embed
 
         has_at_least_one_award = False
@@ -183,7 +183,7 @@ class AdminAction(BaseAction):
             name="Create Voice Channels",
             value=humanize_bool(guild["voice_create"]),
         )
-        embed.color = discord.Color(self.settings.EMBED_COLOR)
+        embed.color = discord.Color(self.settings.INFO_EMBED_COLOR)
         return embed
 
     async def forget_channel(self, channel_str: str) -> None:
@@ -283,7 +283,7 @@ class AdminAction(BaseAction):
         line = award_line(award)
         description = f"{line}\n\nYou can view all awards with the `/set awards` command."
         embed.description = description
-        embed.color = self.settings.EMBED_COLOR
+        embed.color = self.settings.INFO_EMBED_COLOR
         await safe_send_channel(self.interaction, embed=embed, ephemeral=True)
 
     async def award_delete(self, guild_award_id: int) -> None:
@@ -293,7 +293,7 @@ class AdminAction(BaseAction):
         embed.set_author(name="Award deleted!")
         description = "You can view all awards with the `/set awards` command."
         embed.description = description
-        embed.color = self.settings.EMBED_COLOR
+        embed.color = self.settings.INFO_EMBED_COLOR
         await safe_send_channel(self.interaction, embed=embed, ephemeral=True)
 
     async def set_motd(self, message: Optional[str] = None) -> None: